--- a/lean/commands/cloud/live/liquidate.py
+++ b/lean/commands/cloud/live/liquidate.py
@@ -12,13 +12,13 @@
 # limitations under the License.
 
 
-import click
+from click import command, argument
 from lean.click import LeanCommand
 from lean.container import container
 
 
-@click.command(cls=LeanCommand)
-@click.argument("project", type=str)
+@command(cls=LeanCommand)
+@argument("project", type=str)
 def liquidate(project: str) -> None:
     """
     Stops live trading and liquidates existing positions for a certain project.