- Remove, reduce, lazy import modules
- Removing `dependency-injector` as a dependency

- Sanity tested: local live, research, backtest, optimization. Cloud backtest, push, pull, live, optimization. Local data download, generate


Create Project
PR
0s 548ms
0s 540ms
0s 541ms

MASTER
1s 479ms
1s 457ms
1s 458ms